+++ bench/replaceTokenGroups.js
@@ -13,19 +13,19 @@ function benchmark(cb) {
         var replacers = token.createReplacer({
             "\\d+": "###"
         });
-        var res = token.replaceToken(replacers, 'abc ' + Math.round(Math.random() * 1000) + ' def');
+        var res = token.replaceToken(replacers, 'abc ' + Math.round(Math.random() * 1000) + ' def').query;
     })
     suite.add('token replace - numbered groups', function() {
         var replacers = token.createReplacer({
             "(\\d+)": "#$1#"
         });
-        var res = token.replaceToken(replacers, 'abc ' + Math.round(Math.random() * 1000) + ' def');
+        var res = token.replaceToken(replacers, 'abc ' + Math.round(Math.random() * 1000) + ' def').query;
     })
     .add('token replace - named groups', function() {
         var replacers = token.createReplacer({
             "(?<num>\\d+)": "#${num}#"
         });
-        var res = token.replaceToken(replacers, 'abc ' + Math.round(Math.random() * 1000) + ' def');
+        var res = token.replaceToken(replacers, 'abc ' + Math.round(Math.random() * 1000) + ' def').query;
     })
     .on('cycle', function(event) {
         console.log(String(event.target));